requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
astroid==2.4.2
autopep8==1.5.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudinary==1.20.0
coverage==5.0.3
coveralls==1.11.1
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.7
django-braces==1.14.0
django-cors-headers==3.4.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
entrypoints==0.3
gunicorn==20.0.4
idna==2.8
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
oauthlib==3.1.0
Pillow==7.0.0
promise==2.3
psycopg2==2.8.4
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
python-dotenv==0.11.0
python3-openid==3.2.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.12.1
